# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIPAMAS API",
        "description": "Sistem Pelaporan Masyarakat – REST API Documentation. Gunakan JWT Bearer token untuk endpoint yang memerlukan autentikasi.",
        "contact": {
            "name": "SIPAMAS Support",
            "email": "admin@sipamas.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/reports": {
            "get": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Daftar semua laporan",
                "operationId": "42398321972bc3074eef76718e0fd9dc",
                "responses": {
                    "200": {
                        "description": "Daftar laporan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ReportResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Buat laporan baru",
                "operationId": "318de4ef8123b483320dc9dbb45b6a43",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "category_id",
                                    "region",
                                    "location",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Jalan Rusak di RT 05"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "region": {
                                        "type": "string",
                                        "example": "Kec. Ciledug, Kota Tangerang"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Jl. Mawar No. 10"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Jalan berlubang cukup dalam dan membahayakan pengendara"
                                    },
                                    "evidence_photo": {
                                        "description": "Foto bukti (jpg/png, maks 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Laporan berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Laporan berhasil dibuat"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ReportResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reports/{id}": {
            "get": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Detail laporan berdasarkan ID",
                "operationId": "a4c119b5a8302fa439721ad28809a1fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail laporan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ReportResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Laporan tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Hapus laporan milik sendiri",
                "operationId": "a577f2ba73f34e6564524e2163d9bcdf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Laporan berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Laporan berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Tidak diizinkan menghapus laporan milik orang lain"
                    },
                    "404": {
                        "description": "Laporan tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ReportResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Jalan Rusak di RT 05"
                    },
                    "region": {
                        "type": "string",
                        "example": "Kec. Ciledug, Kota Tangerang"
                    },
                    "location": {
                        "type": "string",
                        "example": "Jl. Mawar No. 10"
                    },
                    "description": {
                        "type": "string",
                        "example": "Jalan berlubang cukup dalam dan membahayakan"
                    },
                    "evidence_photo": {
                        "type": "string",
                        "example": "http://localhost:8000/storage/reports/foto.jpg",
                        "nullable": true
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Budi Santoso"
                            }
                        },
                        "type": "object"
                    },
                    "category": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "category_name": {
                                "type": "string",
                                "example": "Infrastruktur"
                            }
                        },
                        "type": "object"
                    },
                    "status": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "status_name": {
                                "type": "string",
                                "example": "Menunggu Verifikasi"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2026-06-07T10:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2026-06-07T10:00:00.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan JWT token dari endpoint /api/login. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}